Fix PR review, e2e file strucutre
