+++ e2e/client/chat-flow/main.fmf
@@ -1,0 +1,16 @@
+summary: Client chat flow - messages, tool approval/denial, reset, SSE replay
+description: |
+    Tests the web client chat workflow including basic messaging,
+    tool approval with real Keylime tool execution, tool denial,
+    conversation reset, and SSE history replay on reconnect.
+contact: Marek Safarik <msafarik@redhat.com>
+component:
+  - keylime-mcp
+framework: shell
+path: /
+test: go test -v -tags=functional ./e2e/client/chat-flow/
+require:
+  - golang
+  - keylime
+duration: 5m
+enabled: true
