+++ e2e/client/model-management/main.fmf
@@ -1,0 +1,16 @@
+summary: Client model management - list, get, set, invalid provider
+description: |
+    Tests the web client model management endpoints including
+    model listing, current model retrieval, model switching,
+    and invalid provider rejection.
+contact: Marek Safarik <msafarik@redhat.com>
+component:
+  - keylime-mcp
+framework: shell
+path: /
+test: go test -v -tags=functional ./e2e/client/model-management/
+require:
+  - golang
+  - keylime
+duration: 5m
+enabled: true
