+++ e2e/server/agent-lifecycle/main.fmf
@@ -1,0 +1,17 @@
+summary: Full agent lifecycle via MCP server
+description: |
+    Tests complete agent lifecycle through MCP tools:
+    registration, enrollment, status, policies, stop,
+    reactivate, update, unenroll, and registrar removal.
+    Runs with data masking enabled.
+contact: Marek Safarik <msafarik@redhat.com>
+component:
+  - keylime-mcp
+framework: shell
+path: /
+test: go test -v -tags=functional ./e2e/server/agent-lifecycle/
+require:
+  - golang
+  - keylime
+duration: 5m
+enabled: true
