+++ e2e/server/policy-management/testdata/test_mb_policy.json
@@ -1,0 +1,7 @@
+{
+  "meta": {
+    "version": 1,
+    "generator": "keylime-mcp-tests"
+  },
+  "events": []
+}
