--- functional-tests/agent-lifecycle/main.fmf
@@ -1,16 +1,0 @@
-summary: Full agent lifecycle via MCP server
-description: |
-    Tests complete agent lifecycle through MCP tools:
-    registration, enrollment, status, policies, stop,
-    reactivate, update, unenroll, and registrar removal.
-    Runs with data masking enabled.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/agent-lifecycle/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
