--- functional-tests/basic-attestation-on-localhost/main.fmf
@@ -1,15 +1,0 @@
-summary: Basic attestation scenario via MCP server
-description: |
-    Tests MCP protocol initialization, tool listing,
-    health check, agent enrollment, status and unenrollment
-    against a running keylime instance.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/basic-attestation-on-localhost/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
