--- functional-tests/client-chat-flow/main.fmf
@@ -1,15 +1,0 @@
-summary: Client chat flow - messages, tool approval/denial, reset, SSE replay
-description: |
-    Tests the web client chat workflow including basic messaging,
-    tool approval with real Keylime tool execution, tool denial,
-    conversation reset, and SSE history replay on reconnect.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/client-chat-flow/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
