--- functional-tests/client-error-handling/main.fmf
@@ -1,14 +1,0 @@
-summary: Client error handling - empty message, no pending tool, index page
-description: |
-    Tests the web client error handling including index page serving,
-    empty message rejection, and tool approve/deny without pending requests.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/client-error-handling/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
