--- functional-tests/error-handling/main.fmf
@@ -1,14 +1,0 @@
-summary: Error handling and edge cases for MCP server
-description: |
-    Tests error responses for invalid inputs, non-existent
-    resources, verifier log queries, and partial service failures.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/error-handling/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
