--- functional-tests/plans/keylime-mcp-main.fmf
@@ -1,39 +1,0 @@
-summary: All functional tests for keylime-mcp (server + client)
-
-context:
-  swtpm: yes
-  agent: rust
-  faked_measured_boot_log: no
-
-discover:
-  - name: keylime-setup
-    how: fmf
-    url: https://github.com/RedHat-SP-Security/keylime-tests.git
-    ref: main
-    test:
-      - /setup/apply_workarounds
-      - /setup/configure_tpm_emulator
-      - /setup/configure_kernel_ima_module/ima_policy_simple
-      - /setup/enable_keylime_debug_messages
-  - name: mcp-server-tests
-    how: fmf
-    test:
-      - /functional-tests/setup-keylime-services
-      - /functional-tests/basic-attestation-on-localhost
-      - /functional-tests/tls-certificate-validation
-      - /functional-tests/policy-management
-      - /functional-tests/agent-lifecycle
-      - /functional-tests/error-handling
-  - name: mcp-client-tests
-    how: fmf
-    test:
-      - /functional-tests/client-chat-flow
-      - /functional-tests/client-model-management
-      - /functional-tests/client-error-handling
-  - name: cleanup
-    how: fmf
-    test:
-      - /functional-tests/cleanup-keylime-services
-
-execute:
-    how: tmt
