--- functional-tests/plans/keylime-mcp-server.fmf
@@ -1,30 +1,0 @@
-summary: Functional tests for keylime-mcp server
-
-context:
-  swtpm: yes
-  agent: rust
-  faked_measured_boot_log: no
-
-discover:
-  - name: keylime-setup
-    how: fmf
-    url: https://github.com/RedHat-SP-Security/keylime-tests.git
-    ref: main
-    test:
-      - /setup/apply_workarounds
-      - /setup/configure_tpm_emulator
-      - /setup/configure_kernel_ima_module/ima_policy_simple
-      - /setup/enable_keylime_debug_messages
-  - name: mcp-server-tests
-    how: fmf
-    test:
-      - /functional-tests/setup-keylime-services
-      - /functional-tests/basic-attestation-on-localhost
-      - /functional-tests/tls-certificate-validation
-      - /functional-tests/policy-management
-      - /functional-tests/agent-lifecycle
-      - /functional-tests/error-handling
-      - /functional-tests/cleanup-keylime-services
-
-execute:
-    how: tmt
