--- functional-tests/policy-management/main.fmf
@@ -1,15 +1,0 @@
-summary: Runtime and measured boot policy management via MCP server
-description: |
-    Tests policy CRUD operations through MCP tools:
-    import, list, get, update, and delete for both
-    runtime policies and measured boot policies.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/policy-management/
-require:
-  - golang
-  - keylime
-duration: 5m
-enabled: true
