--- functional-tests/tls-certificate-validation/main.fmf
@@ -1,16 +1,0 @@
-summary: TLS certificate validation for MCP server
-description: |
-    Tests TLS behavior: valid mTLS, wrong CA certificate,
-    missing client certificate, SNI hostname mismatch,
-    and TLS disabled mode.
-contact: Marek Safarik <msafarik@redhat.com>
-component:
-  - keylime-mcp
-framework: shell
-test: cd ${TMT_TREE} && go test -v -count=1 -timeout=270s -tags=functional ./functional-tests/tls-certificate-validation/
-require:
-  - golang
-  - keylime
-  - openssl
-duration: 5m
-enabled: true
